requirements.txt
# =============================================================================
# Kita.dev Python Dependencies
# =============================================================================

# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# LLM Integration
openai>=1.10.0
litellm>=1.20.0

# GitHub Integration
requests>=2.31.0
PyJWT>=2.8.0
cryptography>=41.0.0

# Configuration
pyyaml>=6.0.0

# Monitoring (optional)
prometheus-client>=0.19.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Code Quality
pylint>=3.0.0
black>=24.1.0
mypy>=1.8.0

# Security
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
